// src/interfaces/CarnetResponse.interface.tsx
// Generated by https://quicktype.io

export interface CarnetResponse {
  count: number;
  page: number;
  totalPages: number;
  data: Carnet[];
}

export interface Carnet {
  status: boolean;
  name: string;
  firstLastName: string;
  secondLastName: string;
  carnet: string;
  reparto?: string;
  deparment?: string;
  floor?: string;
  address: string;
  number: string;
  firstAccross?: string;
  secondAccross?: string;
  municipio: string;
  provincia: string;
  phoneNumber: string;
  user: User;
  createdAt: string;
  updatedAt: string;
  id: string;
}

export interface User {
  name: string;
  id: string;
}